locked_shrinker: Introduction
This class is meant to protect the memory consumption and memory shrinking
critical sections.

It implements the memory::shrinker interface to get triggers for memory
freeing.

It implements lock()/unlock() methods to have a semanticist of standard mutex
and to work with WITH_LOCK() macros.

It prevents the dead-lock situation in the shrinking flow (when shrinking is
caused by the allocation of a new memory for the cache) by using a try_lock()
in the shrinking flow. If there is a lock contention during the shrinking
then it sets the shrinking to the next time the lock() method is called.

Signed-off-by: Vlad Zolotarov <vladz@cloudius-systems.com>